refactor(workspace): reorganize into search/ and tools/ folders
Move search-related files (bm25, search-engine) to search/ subfolder
and tool-related files (tools, tree-formatter) to tools/ subfolder.
Add barrel files for cleaner imports.
