+++ packages/core/src/workspace/search/index.ts
@@ -1,0 +1,2 @@
+export * from './bm25';
+export * from './search-engine';
