+++ packages/core/src/workspace/search/search-engine.test.ts
