+++ packages/core/src/workspace/search/search-engine.ts
@@ -5,12 +5,12 @@
  * semantic (vector), and combined hybrid search across indexed content.
  */
 
-import type { SearchMode, LineRange } from '../artifacts';
-import type { MastraVector, VectorFilter } from '../vector';
+import type { SearchMode, LineRange } from '../../artifacts';
+import type { MastraVector, VectorFilter } from '../../vector';
 
+import { findLineRange } from '../line-utils';
 import { BM25Index, tokenize } from './bm25';
 import type { BM25Config, TokenizeOptions } from './bm25';
-import { findLineRange } from './line-utils';
 
 // Re-export SearchMode for consumers
 export type { SearchMode };
