+++ packages/core/src/workspace/skills/workspace-skills.ts
@@ -11,7 +11,7 @@
 
 import matter from 'gray-matter';
 
-import type { IndexDocument, SearchResult } from '../search-engine';
+import type { IndexDocument, SearchResult } from '../search';
 import { validateSkillMetadata } from './schemas';
 import type { SkillSource as SkillSourceInterface } from './skill-source';
 import { isWritableSource } from './skill-source';
